pyproject.toml
[project]
name = "ralph-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.15.0",
    "openai-agents>=0.6.9",
    "python-dotenv>=1.2.1",
]
